+++ tasks/SPARK/src/main/java/kmer/SparkUtil.java
@@ -1,3 +1,4 @@
+package kmer;
 
 import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaSparkContext;
